+++ glclient/certificate.h
@@ -1,0 +1,1 @@
+extern void checkCertificateExpire(const char *file);
