+++ glclient/client.c
@@ -61,8 +61,7 @@
 #include "utils.h"
 #include "tempdir.h"
 #include "dialogs.h"
-
-#define CERT_EXPIRE_CHECK_MONTHES 2
+#include "certificate.h"
 
 extern void SetSessionTitle(const char *title) {
   if (TITLE(Session)) {
@@ -135,30 +134,6 @@ static void ShowStartupMessage() {
     }
   }
 }
-
-void checkCertificateExpire(const char *file) {
-  FILE *fp;
-  X509 *cert;
-  ASN1_TIME *not_after;
-  int day, sec;
-
-  if ((fp = fopen(file, "rb")) == NULL) {
-    Error("Error open file");
-  }
-  cert = PEM_read_X509(fp, NULL, NULL, NULL);
-
-  not_after = X509_get_notAfter(cert);
-  ASN1_TIME_diff(&day, &sec, NULL, not_after);
-
-  if (day < (CERT_EXPIRE_CHECK_MONTHES * 30)) {
-    time_t t = time(NULL);
-    char s[256];
-    t += day * 24 * 3600 + sec;
-    strftime(s, sizeof(s), _("Certificate Expiration is Approaching.\nexpiration: %Y/%m/%d/ %H:%M:%S"), localtime(&t));
-    MessageDialog(GTK_MESSAGE_INFO, s);
-  }
-}
-
 
 static gboolean StartClient() {
   GLP(Session) = InitProtocol(AuthURI, User, Pass);
